handwriting/requirements.txt
absl-py==0.7.0
aip==1.0.0.5
aircv==1.4.6
airtest==1.0.25
altgraph==0.16.1
anaconda==0.0.1.1
asn1crypto==0.24.0
astor==0.6.2
astroid==2.0.4
atari-py==0.2.6
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
baidu-aip==2.2.8.0
beautifulsoup4==4.7.1
bleach==1.5.0
boto3==1.9.121
botocore==1.12.121
box2d-py==2.3.8
bs4==0.0.1
cairocffi==0.9.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==1.2.1
colorama==0.3.9
comtypes==1.1.7
configparser==3.5.0
cryptography==2.3.1
csvtotable==2.1.1
curses==2.2
cycler==0.10.0
Cython==0.29.12
dask==0.19.4
dataclasses==0.8
decorator==4.3.0
defusedxml==0.6.0
Django==2.1.4
dm-sonnet==1.18
docopt==0.6.2
docutils==0.14
dominate==2.6.0
dytb==0.7.4
easydict==1.8
editdistance==0.5.2
entrypoints==0.3
facebook-wda==0.2.1
flake8==3.5.0
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Script==2.0.6
future==0.17.1
gast==0.2.0
gevent==1.3.6
glfw==1.8.2
google-pasta==0.1.7
graphviz==0.11.1
greenlet==0.4.15
grpcio==1.38.1
grpcio-tools==1.38.1
gym==0.13.1
h5py==2.8.0
hprose==1.4.3
hrpc==1.0.8
html5lib==0.9999999
idna==2.7
imageio==2.5.0
importlib-metadata==4.6.1
imutils==0.5.2
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==0.24
jedi==0.13.3
jieba==0.39
Jinja2==2.10
jmespath==0.9.4
JSAnimation==0.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.5
keras-rl==0.4.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lmdb==0.94
lockfile==0.12.2
locustio==0.9.0
lxml==4.2.4
mahotas==1.4.5
Markdown==2.6.11
MarkupSafe==1.0
mathematics-dataset==1.0.1
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
more-itertools==4.3.0
mpmath==1.1.0
msgpack==0.5.6
namedlist==1.7
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.2.5
notebook==5.7.8
numpy==1.17.0
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
opt-einsum==3.0.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.4.0
pbr==5.1.3
pefile==2019.4.18
pickleshare==0.7.5
pika==1.0.1
Pillow==5.0.0
pipenv==2018.7.1
pipreqs==0.4.10
pluggy==0.7.1
pocoui==1.0.72
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
py==1.6.0
pycodestyle==2.3.1
pycparser==2.18
pydot-ng==2.0.0
pyflakes==1.6.0
pygame==1.9.4
pyglet==1.3.2
Pygments==2.3.1
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
pylint==2.1.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
PyQt4==4.11.4
PyQt5==5.10.1
pyrsistent==0.14.11
pytesseract==0.2.5
pytest==3.8.0
pytest-allure-adaptor==1.7.10
pytest-capturelog==0.7
pytest-html==1.19.0
pytest-metadata==1.7.0
pytest-pythonpath==0.7.3
python-dateutil==2.7.2
python-qt5==0.1.10
pytz==2018.3
PyWavelets==1.1.1
pywin32==223
pywin32-ctypes==0.2.0
pywinauto==0.6.3
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.19.1
Rtree==0.9.3
s3transfer==0.2.0
scikit-image==0.17.2
scikit-learn==0.20.2
scipy==1.0.1
selenium==3.12.0
Send2Trash==1.5.0
simplejson==3.17.3
sip==4.19.8
six==1.11.0
sklearn==0.0
snownlp==0.12.3
soupsieve==1.9.1
sympy==1.4
tb-nightly==1.15.0a20190825
tencentcloud-sdk-python==3.0.185
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.12.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.8.2
tesseract==0.1.3
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019082101
tf-nightly==1.15.0.dev20190821
tflearn==0.3.2
threadpool==1.3.2
thulac==0.2.0
tifffile==2020.9.3
toco==0.0.8
toolz==0.9.0
torch==1.8.0+cu101
torchaudio==0.8.0
torchsummary==1.5.1
torchvision==0.9.0+cu101
tornado==6.0.2
traitlets==4.3.2
trimesh==3.6.43
typed-ast==1.1.0
typing-extensions==3.7.4.3
unicodecsv==0.14.1
UNKNOWN==0.0.0
urllib3==1.23
uuid==1.30
virtualenv==16.0.0
virtualenv-clone==0.3.0
visitor==0.1.3
wcwidth==0.1.7
web.py==0.40.dev0
websocket-client==0.54.0
Werkzeug==0.14.1
wget==3.2
widgetsnbextension==3.4.2
wordcloud==1.5.0
wrapt==1.11.2
wxPython==4.0.1
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.21.0
yarg==0.1.9
yolov4==0.16.0
zipp==3.5.0
